+++ src/features/fan/FanDashboard/components/LivePerformances/style.css.ts
