+++ src/features/fan/services/fan.api.ts
@@ -1,0 +1,6 @@
+import { customAxios } from '@/shared/api';
+
+export const getPerformances = async () => {
+  const { data } = await customAxios.get('/performances');
+  return data;
+};
