+++ src/features/fan/services/fan.key.ts
@@ -1,0 +1,3 @@
+export const fanKeys = {
+  fanLivePerformance: 'fan.query.fanLivePerformance',
+};
